Add files via upload
MODULE: DATA CLEANING AND DATA MANIPULATION 
The data set is made up of:

A master or catalog of products (maestro_productos.csv): contains all the available information related to the products sold by the client and driver.
A master or customer catalog (maestro_clientes.csv): Contains all available information about the stores and distributors that sell the customer's products.
Sellout sales files for two different suppliers (sellout-supplier1.csv and sellout-supplier2.csv): contains sales history and sellout price (from retailer to final consumer) for each product and each store.
sellin sale file (sellin.csv): Contains the sellin price and sales history (from manufacturer to retailer) for each product and store.
 

SellOut is the information of the sales processes that occur from our client's distributors to the final buyers (that is,

when someone buys a product from our client in a supermarket).


SellIn is the information of our client's sales processes towards their distributors (that is, when a supermarket buys a batch of our client's products in order to sell them to final buyers).

The objective is:

Perform a cleanup and adjustment of this data.
Generate a data structure that is easily searchable by the team of data scientists to generate their models.
 
This requires that the data is not only clean, but also consistent with each other and perfectly ordered. An example query that the data scientist team might require is "I need all sellout and sellin records between dates X and Y for products A,B,C, and D in this particular list of stores: T1, T2, T3, T4, T5 and T6" being the variables X,Y, A, B, C, D, T1, T2, T3, T4, T5, and T6.