Add files via upload
MODULE: EDA (EXPLORATORY DATA ANALYSIS).
Question 1: A company needs 1.75mm filament spools for the 3D printing department. They need to measure how close the filament diameter to 1.75mm actually is. Use a gauge tool and sample the diameter five times on the spool: 1.78, 1.75, 1.72, 1.74, 1.77. Find the mean and standard deviation for this set of values.
Question 2: A mobile manufacturer says that the Z-Phone smartphone has a mean consumption life of 42 months with a standard deviation of 8 months. Assuming a normal distribution, what is the probability that a given random Z-Phone will last between 20 and 30 months?
Issue 3: The first customer is not sure that the filament for 3D printers does not have an average diameter of 1.75mm as advertised. 34 samples were measured with a measurement tool. The sample mean is 1.715588 and the sample standard deviation is 0.029252. What is the 99% confidence interval for the mean of the entire spool of filament?
Question 4: The marketing department of a telephone company has started a new advertising campaign and wants to know if it affected sales, which in the past averaged $10,345 per day with a standard deviation of $552. The new ad campaign ran for 45 days and averaged $11,641 in sales. Did the campaign affect sales? Why or why not? (Use a two-tailed test for more reliable significance.)
The objective is:

Answer all questions asked by customers.
Generate a Python code that explains each step taken to resolve the issue.