Add files via upload
MODULE: BASIC VISUALIZATION
To analyze data we need data, a lot of data... and what better than government data, since they are reliable and are usually always very well organized. So let's analyze the data from the referendum on the membership of the United Kingdom to the European Union (Brexit), which took place on June 23, 2016.

The objective of this practice is multiple:

Calculate who is in favor of the EU and who is not.
Calculate the percentages.
Calculate the census, percentage of voters and rejected.
Which counties will be the most in favor of the EU and which will be the least?
See influences of the following variables:
Age.
Education level.
UK people who were not born in the UK.
Salary income.
Jobless people.