Add files via upload
MODULE: ADVANCED VISUALIZATION
Explore the various trends in employee attrition. Some questions to answer What has been the trend over time? Do sociodemographic variables affect?
Identify which are the variables that influence the wear (Attrition) of employees.
Develop a dashboard in which the knowledge acquired is explained and use it to develop a story in which the reasoning used to draw the conclusions is explained.