Add files via upload
MODULE: SUPERVISED MODELS OF MACHINE LEARNING
The practical case consists of finding a supervised model capable of
correctly set of data classify he train For tea we ask that different families of Models this, heart_valid.csv/onu_data.csv
supervised (seen in this unit), where if possible
presents at least 2 different configurations of the same family. for example of SVM
changing the hyperparameter C in the
example, 2 models
configuration.