Add files via upload
MODULE: TEMPORAL SERIES
What is the format of the date variable? (Be careful, the type of variable is not asked). How frequent is it?
Does the time series containing total rainfall follow an additive or multiplicative model?
Next, do the decomposition of the time series. Can you observe a trend or seasonality component?
Is the time series stationary? If the answer is no, make the necessary transformations to make it so.
Create a benchmark model to predict the last 8 periods. Choose an error measure and obtain the error of this estimate
Build an AR or MA model. How does the error of the estimate vary with this new model?
Build an ARIMA or SARIMA model. Analyze the impact on the estimation error.
Use the auto_arima functionality. Does the error of the estimate change?
Build a model with the Prophet library and analyze the change in the error measure.
Finally, which model is the best?