Add files via upload
MODULES: R AND PYTHON PROGRAMMATION.
Time Inc. is a company specialized in providing time-based services to multiple companies and they have detected a need in the market: to create a converter between time units. That is, users provide a number of seconds and want to automatically convert it to minutes, hours, or days.

 

| Development |

As a company developer you are asked to design this functionality with the following requirements:

This functionality must be called through a function called time_converter().
The user is asked for 2 parameters: the number of seconds to convert and the destination unit (minutes, hours or days).
The result should show the original value in seconds and the new value of the equivalence.
The number of seconds cannot be negative and if the user requests its conversion it must show an error message stating that the number to be converted must be positive.
If the user enters a text string in the number of seconds it should also inform the user that this output is not as expected.
Here are some input/output pairs to show the functionality:
time_converter(0, 'minutes') -> 0 seconds equals 0.00 minutes.
time_converter(60, 'minutes') -> 60 seconds equals 1.00 minute.
time_converter(120, 'minutes') -> 60 seconds equals 2.01 minutes.
 

The development of this converter is divided into 3 phases:

Design the flowchart and the corresponding pseudocode.
Generate R code that meets the specified requirements.
Generate the code in Python that meets the specified requirements.