Add files via upload
This data set contains rainfall in Spain for all weather stations. There are some variables broken down by time slot and a variable that is the total sum.

| Development |

You can add new cells if you need to in the template that has been provided.
When delivering the corresponding notebook, use your initials in the name of the file.
It is very important that the notebook is executable and the results are reproducible.
| Requirements for delivery |

Answer the following questions that you will find in the YourInitialsHere_final_exercise.ipynb notebook and upload the file with the answers:

What is the format of the date variable? (Be careful, the type of variable is not asked). How frequent is it?
Does the time series containing total rainfall follow an additive or multiplicative model?
Next, do the decomposition of the time series. Can you observe a trend or seasonality component?
Is the time series stationary? If the answer is no, make the necessary transformations to make it so.
Create a benchmark model to predict the last 8 periods. Choose an error measure and obtain the error of this estimate
Build an AR or MA model. How does the error of the estimate vary with this new model?
Build an ARIMA or SARIMA model. Analyze the impact on the estimation error.
Use the auto_arima functionality. Does the error of the estimate change?
Build a model with the Prophet library and analyze the change in the error measure.
This point does not need to be answered, there is a compatibility problem in Colab.
If it is solved, it will count as an extra point, but if the problem is not solved, it will be evaluated on the remaining 9
Finally, which model is the best?