Add files via upload
A restaurant owner has commissioned you with a tool that automatically sends emails when his restaurant receives a negative review. Your objective is to train a model that knows how to detect how positive or negative it is from the text of a review.

We are going to work with the dataset restaurant_ratings.csv (also available in the Module 5 resources) which has 4 columns:

Name: the name of the restaurant
Rating: the number of stars given by the user.
Review: the review written by the user.
Bad: 0 if the review is considered good (if its rating is equal to or greater than 2.5), 1 if it is considered bad (if its rating is less than 2.5).
You have to train a classification model using the "bad" column as target. Please remove the "name" and "rating" columns. The first does not contribute anything useful. And the second is that we don't know it when we receive a review, since in the proposed scenario it is assumed that we only receive the text of it.

As a reminder, the main steps to take are:

Exploration of the data (see if there are nulls, eliminate them, etc.).
Preprocessing and cleaning of texts
Vectorization of the texts (with your favorite method, or you can try several!)
Model training (you can work with any type of modeling).
Assessment. Remember here also that the classification is multiclass, so you have to use metrics that are suitable for this type of classification.

| Development |
Although the goal is for the model to give us the best possible metrics, for this exercise I am going to focus more on the steps you have taken and why. I don't care so much if you get to a very good model or not, but everything you do along the way.